chore: moves datastore types/interface
