+++ packages/amplify/amplify_flutter/example/lib/main.dart
@@ -14,6 +14,7 @@
  */
 
 import 'package:amplify_flutter/amplify_flutter.dart';
+import 'package:aws_common/aws_common.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
