+++ packages/amplify/amplify_flutter/lib/src/amplify_impl.dart
@@ -17,7 +17,6 @@ import 'dart:async';
 import 'dart:convert';
 
 import 'package:amplify_core/amplify_core.dart';
-import 'package:amplify_datastore_plugin_interface/amplify_datastore_plugin_interface.dart';
 import 'package:aws_common/aws_common.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
@@ -43,7 +42,7 @@ class AmplifyClass extends PlatformInterface {
   final StorageCategory Storage = StorageCategory();
 
   /// The DataStore category.
-  final DataStoreCategory DataStore = const DataStoreCategory();
+  final DataStoreCategory DataStore = DataStoreCategory();
 
   /// The API category.
   final APICategory API = APICategory();
