+++ packages/amplify_core/lib/src/category/amplify_categories.dart
@@ -19,10 +19,7 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:amplify_core/amplify_core.dart';
-import 'package:amplify_core/src/types/plugin/amplify_analytics_plugin_interface.dart';
-import 'package:amplify_core/src/types/plugin/amplify_auth_plugin_interface.dart';
-import 'package:amplify_core/src/types/plugin/amplify_plugin_interface.dart';
-import 'package:amplify_core/src/types/plugin/amplify_storage_plugin_interface.dart';
+import 'package:amplify_core/src/types/plugin/amplify_datastore_plugin_interface.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter/services.dart';
 import 'package:meta/meta.dart';
