+++ packages/amplify_core/lib/src/types/datastore/exception/datstore_exception.dart
