+++ packages/amplify_core/lib/src/types/datastore/hub/datstore_hub_event.dart
