+++ packages/amplify_core/lib/src/types/datastore/hub/hub_event_element.dart
@@ -13,9 +13,9 @@
  * permissions and limitations under the License.
  */
 
-import 'package:amplify_datastore_plugin_interface/amplify_datastore_plugin_interface.dart';
+import 'package:amplify_core/amplify_core.dart';
 
-part 'HubEventElementWithMetadata.dart';
+part 'hub_event_element_with_metadata.dart';
 
 /// The model associated with a DataStore `outboxMutationEnqueued` or
 /// `outboxMutationProcessed` Hub event.
