+++ packages/amplify_core/lib/src/types/datastore/hub/model_synced_event.dart
