+++ packages/amplify_core/lib/src/types/datastore/hub/network_status_event.dart
