+++ packages/amplify_core/lib/src/types/datastore/hub/outbox_status_event.dart
