+++ packages/amplify_core/lib/src/types/datastore/hub/sync_queries_started_event.dart
