+++ packages/amplify_core/lib/src/types/datastore/models/auth_mode_strategy.dart
