+++ packages/amplify_core/lib/src/types/datastore/models/observe_query_throttle_options.dart
