+++ packages/amplify_core/lib/src/types/datastore/models/query_snapshot.dart
