+++ packages/amplify_core/lib/src/types/datastore/models/subscription_event.dart
