+++ packages/amplify_core/lib/src/types/plugin/amplify_plugin_interface.dart
@@ -30,7 +30,7 @@ abstract class AmplifyPluginInterface {
   Future<void> addPlugin() async {}
 
   /// Configures the plugin using the registered [config].
-  Future<void> configure({AmplifyConfig? config}) async {}
+  Future<void> configure({String? config}) async {}
 
   /// Resets the plugin by removing all traces of it from the device.
   @visibleForTesting
