+++ packages/amplify_core/lib/src/util/uuid.dart
