+++ packages/amplify_core/pubspec.yaml
@@ -16,6 +16,7 @@ dependencies:
   date_time_format: ^2.0.1
   json_annotation: ^4.3.0
   meta: ^1.3.0
+  uuid: ^3.0.1
 
 dev_dependencies:
   amplify_lints: ^1.0.0
