+++ packages/amplify_core/test/datstore/test/amplify_custom_type_schema_test.dart
@@ -13,8 +13,8 @@
  * permissions and limitations under the License.
  */
 
+import 'package:amplify_core/amplify_core.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:amplify_datastore_plugin_interface/amplify_datastore_plugin_interface.dart';
 
 import 'testData/ModelProvider.dart';
