+++ packages/amplify_core/test/datstore/test/sorted_list_test.dart
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-import 'package:amplify_datastore_plugin_interface/src/types/models/sorted_list.dart';
+import 'package:amplify_core/src/types/datastore/models/sorted_list.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
