+++ packages/amplify_core/test/datstore/test/testData/Comment.dart
@@ -1,0 +1,227 @@
+/*
+* Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+*
+* Licensed under the Apache License, Version 2.0 (the "License").
+* You may not use this file except in compliance with the License.
+* A copy of the License is located at
+*
+*  http://aws.amazon.com/apache2.0
+*
+* or in the "license" file accompanying this file. This file is distributed
+* on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+* express or implied. See the License for the specific language governing
+* permissions and limitations under the License.
+*/
+
+// NOTE: This file is generated and may not follow lint rules defined in your app
+// Generated files can be excluded from analysis in analysis_options.yaml
+// For more info, see: https://dart.dev/guides/language/analysis-options#excluding-code-from-analysis
+
+// ignore_for_file: public_member_api_docs, file_names, unnecessary_new, prefer_if_null_operators, prefer_const_constructors, slash_for_doc_comments, annotate_overrides, non_constant_identifier_names, unnecessary_string_interpolations, prefer_adjacent_string_concatenation, unnecessary_const, dead_code, implicit_dynamic_map_literal, implicit_dynamic_parameter, implicit_dynamic_type
+
+import 'ModelProvider.dart';
+import 'package:amplify_core/amplify_core.dart';
+import 'package:flutter/foundation.dart';
+
+/** This is an auto generated class representing the Comment type in your schema. */
+@immutable
+class Comment extends Model {
+  static const classType = const _CommentModelType();
+  final String id;
+  final String? _postID;
+  final Post? _post;
+  final String? _content;
+  final TemporalDateTime? _createdAt;
+  final TemporalDateTime? _updatedAt;
+
+  @override
+  getInstanceType() => classType;
+
+  @override
+  String getId() {
+    return id;
+  }
+
+  String get postID {
+    try {
+      return _postID!;
+    } catch (e) {
+      throw new AmplifyCodeGenModelException(
+          AmplifyExceptionMessages
+              .codeGenRequiredFieldForceCastExceptionMessage,
+          recoverySuggestion: AmplifyExceptionMessages
+              .codeGenRequiredFieldForceCastRecoverySuggestion,
+          underlyingException: e.toString());
+    }
+  }
+
+  Post? get post {
+    return _post;
+  }
+
+  String get content {
+    try {
+      return _content!;
+    } catch (e) {
+      throw new AmplifyCodeGenModelException(
+          AmplifyExceptionMessages
+              .codeGenRequiredFieldForceCastExceptionMessage,
+          recoverySuggestion: AmplifyExceptionMessages
+              .codeGenRequiredFieldForceCastRecoverySuggestion,
+          underlyingException: e.toString());
+    }
+  }
+
+  TemporalDateTime? get createdAt {
+    return _createdAt;
+  }
+
+  TemporalDateTime? get updatedAt {
+    return _updatedAt;
+  }
+
+  const Comment._internal(
+      {required this.id,
+      required postID,
+      post,
+      required content,
+      createdAt,
+      updatedAt})
+      : _postID = postID,
+        _post = post,
+        _content = content,
+        _createdAt = createdAt,
+        _updatedAt = updatedAt;
+
+  factory Comment(
+      {String? id,
+      required String postID,
+      Post? post,
+      required String content}) {
+    return Comment._internal(
+        id: id == null ? UUID.getUUID() : id,
+        postID: postID,
+        post: post,
+        content: content);
+  }
+
+  bool equals(Object other) {
+    return this == other;
+  }
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(other, this)) return true;
+    return other is Comment &&
+        id == other.id &&
+        _postID == other._postID &&
+        _post == other._post &&
+        _content == other._content;
+  }
+
+  @override
+  int get hashCode => toString().hashCode;
+
+  @override
+  String toString() {
+    var buffer = new StringBuffer();
+
+    buffer.write("Comment {");
+    buffer.write("id=" + "$id" + ", ");
+    buffer.write("postID=" + "$_postID" + ", ");
+    buffer.write("post=" + (_post != null ? _post!.toString() : "null") + ", ");
+    buffer.write("content=" + "$_content" + ", ");
+    buffer.write("createdAt=" +
+        (_createdAt != null ? _createdAt!.format() : "null") +
+        ", ");
+    buffer.write(
+        "updatedAt=" + (_updatedAt != null ? _updatedAt!.format() : "null"));
+    buffer.write("}");
+
+    return buffer.toString();
+  }
+
+  Comment copyWith({String? id, String? postID, Post? post, String? content}) {
+    return Comment._internal(
+        id: id ?? this.id,
+        postID: postID ?? this.postID,
+        post: post ?? this.post,
+        content: content ?? this.content);
+  }
+
+  Comment.fromJson(Map<String, dynamic> json)
+      : id = json['id'],
+        _postID = json['postID'],
+        _post = json['post']?['serializedData'] != null
+            ? Post.fromJson(
+                new Map<String, dynamic>.from(json['post']['serializedData']))
+            : null,
+        _content = json['content'],
+        _createdAt = json['createdAt'] != null
+            ? TemporalDateTime.fromString(json['createdAt'])
+            : null,
+        _updatedAt = json['updatedAt'] != null
+            ? TemporalDateTime.fromString(json['updatedAt'])
+            : null;
+
+  Map<String, dynamic> toJson() => {
+        'id': id,
+        'postID': _postID,
+        'post': _post?.toJson(),
+        'content': _content,
+        'createdAt': _createdAt?.format(),
+        'updatedAt': _updatedAt?.format()
+      };
+
+  static final QueryField ID = QueryField(fieldName: "comment.id");
+  static final QueryField POSTID = QueryField(fieldName: "postID");
+  static final QueryField POST = QueryField(
+      fieldName: "post",
+      fieldType: ModelFieldType(ModelFieldTypeEnum.model,
+          ofModelName: (Post).toString()));
+  static final QueryField CONTENT = QueryField(fieldName: "content");
+  static var schema =
+      Model.defineSchema(define: (ModelSchemaDefinition modelSchemaDefinition) {
+    modelSchemaDefinition.name = "Comment";
+    modelSchemaDefinition.pluralName = "Comments";
+
+    modelSchemaDefinition.addField(ModelFieldDefinition.id());
+
+    modelSchemaDefinition.addField(ModelFieldDefinition.field(
+        key: Comment.POSTID,
+        isRequired: true,
+        ofType: ModelFieldType(ModelFieldTypeEnum.string)));
+
+    modelSchemaDefinition.addField(ModelFieldDefinition.belongsTo(
+        key: Comment.POST,
+        isRequired: false,
+        targetName: "postCommentsId",
+        ofModelName: (Post).toString()));
+
+    modelSchemaDefinition.addField(ModelFieldDefinition.field(
+        key: Comment.CONTENT,
+        isRequired: true,
+        ofType: ModelFieldType(ModelFieldTypeEnum.string)));
+
+    modelSchemaDefinition.addField(ModelFieldDefinition.nonQueryField(
+        fieldName: 'createdAt',
+        isRequired: false,
+        isReadOnly: true,
+        ofType: ModelFieldType(ModelFieldTypeEnum.dateTime)));
+
+    modelSchemaDefinition.addField(ModelFieldDefinition.nonQueryField(
+        fieldName: 'updatedAt',
+        isRequired: false,
+        isReadOnly: true,
+        ofType: ModelFieldType(ModelFieldTypeEnum.dateTime)));
+  });
+}
+
+class _CommentModelType extends ModelType<Comment> {
+  const _CommentModelType();
+
+  @override
+  Comment fromJson(Map<String, dynamic> jsonData) {
+    return Comment.fromJson(jsonData);
+  }
+}
