+++ packages/amplify_core/test/datstore/test/testData/ModelProvider.dart
@@ -13,18 +13,21 @@
 * permissions and limitations under the License.
 */
 
-// ignore_for_file: public_member_api_docs
+// NOTE: This file is generated and may not follow lint rules defined in your app
+// Generated files can be excluded from analysis in analysis_options.yaml
+// For more info, see: https://dart.dev/guides/language/analysis-options#excluding-code-from-analysis
 
-import 'package:amplify_datastore_plugin_interface/amplify_datastore_plugin_interface.dart';
+// ignore_for_file: public_member_api_docs, file_names, unnecessary_new, prefer_if_null_operators, prefer_const_constructors, slash_for_doc_comments, annotate_overrides, non_constant_identifier_names, unnecessary_string_interpolations, prefer_adjacent_string_concatenation, unnecessary_const, dead_code
 
-import 'Address.dart';
+import 'package:amplify_core/amplify_core.dart';
 import 'Blog.dart';
 import 'Comment.dart';
+import 'Post.dart';
+import 'PostAuthComplex.dart';
+import 'Address.dart';
 import 'Contact.dart';
 import 'Person.dart';
 import 'Phone.dart';
-import 'Post.dart';
-import 'PostAuthComplex.dart';
 
 export 'Address.dart';
 export 'Blog.dart';
@@ -37,27 +40,17 @@ export 'PostAuthComplex.dart';
 
 class ModelProvider implements ModelProviderInterface {
   @override
-  String version = "836a1f8cc4a4adc493165acc7168289e";
+  String version = "3dbdb02012ee7733ca199fde1f1b8c06";
   @override
-  List<ModelSchema> modelSchemas = [
-    Blog.schema,
-    Comment.schema,
-    Post.schema,
-    PostAuthComplex.schema
-  ];
+  List<ModelSchema> modelSchemas = [Blog.schema, Comment.schema, Post.schema, PostAuthComplex.schema];
   static final ModelProvider _instance = ModelProvider();
   @override
-  List<ModelSchema> customTypeSchemas = [
-    Address.schema,
-    Contact.schema,
-    Person.schema,
-    Phone.schema
-  ];
+  List<ModelSchema> customTypeSchemas = [Address.schema, Contact.schema, Person.schema, Phone.schema];
 
   static ModelProvider get instance => _instance;
-
+  
   ModelType getModelTypeByModelName(String modelName) {
-    switch (modelName) {
+    switch(modelName) {
       case "Blog":
         return Blog.classType;
       case "Comment":
@@ -67,9 +60,7 @@ class ModelProvider implements ModelProviderInterface {
       case "PostAuthComplex":
         return PostAuthComplex.classType;
       default:
-        throw Exception(
-            "Failed to find model in model provider for model name: " +
-                modelName);
+        throw Exception("Failed to find model in model provider for model name: " + modelName);
     }
   }
 }
