+++ packages/amplify_core/test/datstore/test/testData/Person.dart
@@ -13,40 +13,32 @@
 * permissions and limitations under the License.
 */
 
-// ignore_for_file: public_member_api_docs
+// NOTE: This file is generated and may not follow lint rules defined in your app
+// Generated files can be excluded from analysis in analysis_options.yaml
+// For more info, see: https://dart.dev/guides/language/analysis-options#excluding-code-from-analysis
+
+// ignore_for_file: public_member_api_docs, file_names, unnecessary_new, prefer_if_null_operators, prefer_const_constructors, slash_for_doc_comments, annotate_overrides, non_constant_identifier_names, unnecessary_string_interpolations, prefer_adjacent_string_concatenation, unnecessary_const, dead_code, implicit_dynamic_map_literal, implicit_dynamic_parameter, implicit_dynamic_type
 
 import 'ModelProvider.dart';
-import 'package:amplify_datastore_plugin_interface/amplify_datastore_plugin_interface.dart';
+import 'package:amplify_core/amplify_core.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 
 /** This is an auto generated class representing the Person type in your schema. */
 @immutable
-class Person extends Model {
-  static const classType = const _PersonModelType();
-  final String id;
+class Person {
   final String? _name;
   final List<Address>? _propertiesAddresses;
   final Contact? _contact;
-  final TemporalDateTime? _createdAt;
-  final TemporalDateTime? _updatedAt;
-
-  @override
-  getInstanceType() => classType;
-
-  @override
-  String getId() {
-    return id;
-  }
 
   String get name {
     try {
       return _name!;
     } catch (e) {
-      throw new DataStoreException(
-          DataStoreExceptionMessages
+      throw new AmplifyCodeGenModelException(
+          AmplifyExceptionMessages
               .codeGenRequiredFieldForceCastExceptionMessage,
-          recoverySuggestion: DataStoreExceptionMessages
+          recoverySuggestion: AmplifyExceptionMessages
               .codeGenRequiredFieldForceCastRecoverySuggestion,
           underlyingException: e.toString());
     }
@@ -60,43 +52,25 @@ class Person extends Model {
     try {
       return _contact!;
     } catch (e) {
-      throw new DataStoreException(
-          DataStoreExceptionMessages
+      throw new AmplifyCodeGenModelException(
+          AmplifyExceptionMessages
               .codeGenRequiredFieldForceCastExceptionMessage,
-          recoverySuggestion: DataStoreExceptionMessages
+          recoverySuggestion: AmplifyExceptionMessages
               .codeGenRequiredFieldForceCastRecoverySuggestion,
           underlyingException: e.toString());
     }
   }
 
-  TemporalDateTime? get createdAt {
-    return _createdAt;
-  }
-
-  TemporalDateTime? get updatedAt {
-    return _updatedAt;
-  }
-
-  const Person._internal(
-      {required this.id,
-      required name,
-      propertiesAddresses,
-      required contact,
-      createdAt,
-      updatedAt})
+  const Person._internal({required name, propertiesAddresses, required contact})
       : _name = name,
         _propertiesAddresses = propertiesAddresses,
-        _contact = contact,
-        _createdAt = createdAt,
-        _updatedAt = updatedAt;
+        _contact = contact;
 
   factory Person(
-      {String? id,
-      required String name,
+      {required String name,
       List<Address>? propertiesAddresses,
       required Contact contact}) {
     return Person._internal(
-        id: id == null ? UUID.getUUID() : id,
         name: name,
         propertiesAddresses: propertiesAddresses != null
             ? List<Address>.unmodifiable(propertiesAddresses)
@@ -112,7 +86,6 @@ class Person extends Model {
   bool operator ==(Object other) {
     if (identical(other, this)) return true;
     return other is Person &&
-        id == other.id &&
         _name == other._name &&
         DeepCollectionEquality()
             .equals(_propertiesAddresses, other._propertiesAddresses) &&
@@ -127,40 +100,29 @@ class Person extends Model {
     var buffer = new StringBuffer();
 
     buffer.write("Person {");
-    buffer.write("id=" + "$id" + ", ");
     buffer.write("name=" + "$_name" + ", ");
     buffer.write("propertiesAddresses=" +
         (_propertiesAddresses != null
             ? _propertiesAddresses!.toString()
             : "null") +
         ", ");
-    buffer.write(
-        "contact=" + (_contact != null ? _contact!.toString() : "null") + ", ");
-    buffer.write("createdAt=" +
-        (_createdAt != null ? _createdAt!.format() : "null") +
-        ", ");
-    buffer.write(
-        "updatedAt=" + (_updatedAt != null ? _updatedAt!.format() : "null"));
+    buffer
+        .write("contact=" + (_contact != null ? _contact!.toString() : "null"));
     buffer.write("}");
 
     return buffer.toString();
   }
 
   Person copyWith(
-      {String? id,
-      String? name,
-      List<Address>? propertiesAddresses,
-      Contact? contact}) {
+      {String? name, List<Address>? propertiesAddresses, Contact? contact}) {
     return Person._internal(
-        id: id ?? this.id,
         name: name ?? this.name,
         propertiesAddresses: propertiesAddresses ?? this.propertiesAddresses,
         contact: contact ?? this.contact);
   }
 
   Person.fromJson(Map<String, dynamic> json)
-      : id = json['id'],
-        _name = json['name'],
+      : _name = json['name'],
         _propertiesAddresses = json['propertiesAddresses'] is List
             ? (json['propertiesAddresses'] as List)
                 .where((e) => e != null)
@@ -171,38 +133,22 @@ class Person extends Model {
         _contact = json['contact']?['serializedData'] != null
             ? Contact.fromJson(new Map<String, dynamic>.from(
                 json['contact']['serializedData']))
-            : null,
-        _createdAt = json['createdAt'] != null
-            ? TemporalDateTime.fromString(json['createdAt'])
-            : null,
-        _updatedAt = json['updatedAt'] != null
-            ? TemporalDateTime.fromString(json['updatedAt'])
             : null;
 
   Map<String, dynamic> toJson() => {
-        'id': id,
         'name': _name,
         'propertiesAddresses':
             _propertiesAddresses?.map((Address? e) => e?.toJson()).toList(),
-        'contact': _contact?.toJson(),
-        'createdAt': _createdAt?.format(),
-        'updatedAt': _updatedAt?.format()
+        'contact': _contact?.toJson()
       };
 
-  static final QueryField ID = QueryField(fieldName: "person.id");
-  static final QueryField NAME = QueryField(fieldName: "name");
-  static final QueryField PROPERTIESADDRESSES =
-      QueryField(fieldName: "propertiesAddresses");
-  static final QueryField CONTACT = QueryField(fieldName: "contact");
   static var schema =
       Model.defineSchema(define: (ModelSchemaDefinition modelSchemaDefinition) {
     modelSchemaDefinition.name = "Person";
     modelSchemaDefinition.pluralName = "People";
 
-    modelSchemaDefinition.addField(ModelFieldDefinition.id());
-
-    modelSchemaDefinition.addField(ModelFieldDefinition.field(
-        key: Person.NAME,
+    modelSchemaDefinition.addField(ModelFieldDefinition.customTypeField(
+        fieldName: 'name',
         isRequired: true,
         ofType: ModelFieldType(ModelFieldTypeEnum.string)));
 
@@ -218,26 +164,5 @@ class Person extends Model {
         isRequired: true,
         ofType: ModelFieldType(ModelFieldTypeEnum.embedded,
             ofCustomTypeName: 'Contact')));
-
-    modelSchemaDefinition.addField(ModelFieldDefinition.nonQueryField(
-        fieldName: 'createdAt',
-        isRequired: false,
-        isReadOnly: true,
-        ofType: ModelFieldType(ModelFieldTypeEnum.dateTime)));
-
-    modelSchemaDefinition.addField(ModelFieldDefinition.nonQueryField(
-        fieldName: 'updatedAt',
-        isRequired: false,
-        isReadOnly: true,
-        ofType: ModelFieldType(ModelFieldTypeEnum.dateTime)));
   });
-}
-
-class _PersonModelType extends ModelType<Person> {
-  const _PersonModelType();
-
-  @override
-  Person fromJson(Map<String, dynamic> jsonData) {
-    return Person.fromJson(jsonData);
-  }
 }
