+++ packages/aws_common/lib/aws_common.dart
@@ -34,4 +34,3 @@ export 'src/util/equatable.dart';
 export 'src/util/json.dart';
 export 'src/util/print.dart';
 export 'src/util/serializable.dart';
-export 'src/util/uuid.dart';
